refactor: use a consts file
